requirements.txt
asgiref==3.8.1
./diploma_frontend-0.6.tar.gz
Django==5.2.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
python-dotenv==1.1.0
pillow==11.2.1
sqlparse==0.5.3
gunicorn==23.0.0
